[stm] [kernel] Remove support for `vio`, `vio -> vo`
This is an experiment to assess what is the exact complexity of
supporting `vio -> vo` as opposed to a simpler mode where opaque
proofs are just not checked.

The patch is not intended to be merged in any form.

I did the removal in a very quick way, more refactoring may be
possible.

A good side-effect of this PR is that a few kernel interfaces stop
depending on `Future.t`, which IMHO is a good thing as they should
not. The inner workings of `library` are also greatly simplified.

Also, an ugly "hook" in `Opaqueproof` goes away, this seemed a bit
worrying as the kernel should not have such hooks set from the upper
layers. [note one extra hook do remain]