[kernel] Move lazy object loading to the kernel.
This fixes a key problem in regards of making the kernel
self-contained as it removes the "indirect_opaque_accessor" hook, used
to perform lazy-loading of opaque data.

Indeed, that part was a key component of the trusted base, however it
was living in library and made available to the kernel thru a "hook".

Instead, we add to the environment information about where the opaque
data lives and fetch it under demand. We need to be careful as to keep
the environments marshallable thus we add local cache to
`Opaqueproofs`.

This refactors quite a bit of code, and IMO opens the door to further
improvements in `Opaqueproofs`.

Based on #8783, but I think it could be made independent.