[dune] [gitlab] Test OCaml trunk.
We add a job testing the build of Coq with OCaml 4.08 [AKA `trunk`]

CoqIDE is not supported in 4.08 due to missing `lablgtk`, also `oUnit`
cannot be currently installed, thus we have to add a switch to the
test suite to disable `unit-tests`.

Many deprecation warnings happened in 4.08 so we use the `release`
profile to make them not fatal. Using a 4.08 build profile would be an
option too.