[dune] Add flag to drop dependency of `.vo` files on coqtop.
In #8900 there was a lot of discussion about having a fast-path that
skips `.vo` integrity checks by allowing to rebuild `coqtop` but not
the `.vo` that depend on it.

We thus add a flag for such adventurous users so they can manually
control the build. Note that using this may create all kind of havoc,
for example if you haven't build `coqtop` Dune could well pick the one
in your PATH, etc...

This can be improved, but in the end it will put a lot of burden on
the user to check for consistency.