[Docker] Update COMPILER_EDGE: 4.07.0 → 4.07.1
Also update CAMLP5_VER_EDGE: 7.06 → 7.06.10-g84ce6cc4

This is to avoid an OCaml bug that occurs when compiling the OCaml code
extracted from part of a patched version of CompCert.

~~~
File "extraction/Parser.ml", line 12375, characters 19-29:
Error: Constraints are not satisfied in this type.
       Type initstate' should be an instance of initstate'
~~~

This compiler issue only appears with OCaml 4.07.0 (neither with 4.06
nor with 4.07.1); hence this update.