Remove int31_syntax_plugin
The error messages have changed a bit, and we now support parsing
negative numerals as int31.  (If we wanted to not do this, that would be
pretty easy, we'd just need to write a version of `phi_inv` that
returned `option int31` and rejected negative `Z` inputs.)

c.f. https://github.com/coq/www/pull/80#issuecomment-436409783