[topfmt] Add phase attribute for toplevel printing.
This is localized version of #8833, but instead of adding a phase
attribute which, as pointed by @gares has some problematic semantics,
we add a local one to the toplevel functions.

This moves the imperative part of the API to a better-delimited scope
and allows to progress with the separation of the interactive and
compilation API.

Note that still quite a few issues do remain in the "Feedback" path,
for example, idetop and other feedback clients cannot get a hold of
the feedback early enough as to direct init messages to the IDE part.

This is for example a serious issue of the API that shall be treated
separately.