coq_makefile: Fix ocamldep ignoring mlg files
If you have file1.mlg and file2.ml, with file2 depending on file1,
ocamldep was before generating file1.ml so wouldn't generate
[file2.cmx: file1.cmx] (ocamldep is silent on non-found dependencies).

This has been causing nondeterministic failures in quickchick
recently.

I guess it didn't come up in the past because ml4 files tend to be at
the end of the dependency chain.