[STM] Simplify cache logic and interface
Before this patch, functions fetching states from the cache were
expecting clients to check if the state was present and had the expected
shape (success state,...). Querying whether a state was cached could
itself trigger caching of states.

We simplify the interface by providing essentially two functions: one
for installing a state from the cache, and one for getting a state (to
be sent to a worker). The logic related to the current state is also
encapsulated in this interface, rather than carried out by the clients.