CoqIDE: Ensuring that gtk is initialized before other inits done in ideutils.ml.
This seems fragile: does it depend on the order files are loaded? (It
was working for gtk2 when gtk initialization was in coqide_main.ml but
it does not work anymore for CoqIDE built on gtk3).

Eventually, it might be needed to centralize all initialization side
effects in one place.