[dune] Shim for starting `coqtop/coqide` with minimal config.
As requested by Gaëtan Gilbert, we add shims

- `dev/shim/coqtop-prelude`
- `dev/shim/coqide-prelude`

that will build and start `coqtop` and `coqide` with just the prelude
loaded properly.

`dune exec dev/shim/coqtop-prelude` will build and execute this shim,
equivalent to doing `make states && bin/coqtop` under the old model.

This PR is just a bit of "a hack" until proper support for Coq
libraries arrives to Dune, however there is nothing wrong with it.

In particular, we must bootstrap `coq.plugins.ltac` as Dune needs to
compute the full installation path to allow `%{bin:foo}` in deps,
[this is a kind of shortcoming of the current implementation, and the
error message is just terrible]

We cannot depend on installed `.vo` files without doing a gross hack
[including them inside an ml lib] so for now we just depend on their
non-installed forms. Using `%{bin}` is good enough for the shims who
would like to locate binaries using `PATH`.

The long term plan (for now) is to have a command similar to `dune
utop $dir`, `dune coqtop $dir`, which would spawn a proper Coq shell
with the corresponding libraries on the path.

This will work for `dir=stdlib/Init/` for example, or for any other
combination.