[gramlib] [pcoq] Don't delete levels on rule removal.
As requested by @herbelin, this is an experiment to keep the level
when removing the last rule.

This should avoid the need to workaround the level deletion on rule
removal by re-creating it.

The changes in gramext are likely wrong as we may want to modify the
level a bit more I think.