[contributing] [changes] Improve accuracy of credits.
It has been a while that I am not very happy with our current system
to record changes and credits. I see several problems:

- imprecise credit for many features, for example, reviewers are not
  acknowledged [and are sometimes crucial for the design, etc...],
- dissociation of credits and change entries,
- lack of categorization / systematization.

This is indeed a hard problem to solve (cc: #8180) ; in this PR I have
a very modest proposal improve accuracy and traceability by adapting
the model used in OCaml.

In this model, every change entry includes the name of the author, the
PR numbers, and a list of additional acknowledgements, where revievers
are usually mandatory.

I have filled in the information for a few entries I am responsible
of; please let me know what you think.