Don't bundle Z.div_mod_quot_rem into zify
Alas, I have not had time to work on imrpoving the performance of nia,
and there has been a request to include this tactic (which is useful on
its own) without bundling it into `zify`.  So that is what we do here.

I leave the definition of it in `PreOmega` in case we want to eventually
include it in `zify`/`nia`.