[vernac] Remove obsolete Existential and Grab Existential variables.
This is a proposal to eventually remove the commands `Existential` and
`Grab Existential`. To my understanding these commands are supposed to
be deprecated and scheduled to removal, but not 100% sure so we would
need to confirm.

This would of course need a deprecation phase, let's see the impact on
CI first.