Adding proper support for generic applicative notations.
This means that we support notations of the form

  Notation "f # x" := (f x) (at level 10).

in arbitrary positions of an application and with an explicit priority
given to non generic applications. See Notations4.v for details.

In passing, we simplify and centralize in notation.ml the detection of
those lonely notations which may hide notations which normally would
not require a delimiter but which require one because of the masking.