[ci] [appveyor] Temporally disable test suite on Appveyor.
Hopefully we will re-enable it back soon, I am preparing a refactoring
that should make it more robust w.r.t paths and changes on Windows
which will enable to use a faster build system.

But now it is timing out 100% of the times [due to #8655] so it is not
useful.