[dune] Fix states target to depend on installed libraries.
Until proper Coq support comes to Dune, this is a workaround and an
alternative to #9047 so `dune build @states` will build a working
`coqtop`.

In the end the shim path seems better to me, and this is the way `dune
coqtop` will work [and it will allow to obtain a context with
arbitrary libraries], however this is an interesting option to
consider.