Do not rename implicit argument names.
Automatic implicit arguments get the name given by the user,
if they are inferable and no previous binder used the same name.
Otherwise, they are marked uninferable, because the current system
requires all implicit arguments to have a name.

This changes the implicit status of the motive in `ex_ind`, for example,
because the motive is rigid but is not given a name (the most obvious
name, `P`, also conflicts with the second argument of `ex`).
The hope is that this sort of change is unusual enough that CI has few problems.

Allowing anonymous implicit arguments would be a step forward, fixing
this case and others, but is more disruptive and could introduce other
incompatibilities.

If necessary, a compatibility option could be introduced which uses
the old naming algorithm in `find_displayed_name_in`, at the expense
of keeping the `all` parameter around for a while.