Add badges linking to a selection of up-to-date Coq packages.
The badges displaying the latest packaged version are provided by Repology.
The goal is to put forward the packages that are well maintained (updated
quickly after a new release) and thus recommended.

The history of package updates can be found at:
https://repology.org/metapackage/coq/history

opam is missing because Repology is not currently tracking opam, but this
could be fixed (Repology does not only track Linux-distributions-like
package repositories, it also includes programming language package
repositories such as CPAN, CRAN, Hackage, RubyGems, and Stackage, thus
opam could be added as well).