Track polymorphism outside the kernel
Needed to hint at what should be done for schemes (this is pretty
hacky though)
BTW we get an anomaly in Term_typing.skip_trusted_effect if we make
all schemes polymorphic.

Consider
~~~coq
Polymorphic Inductive foo : Set := bar.
~~~
The foo_rect eliminator is expected to be polymorphic, this is done by
having the same polymorphism as foo. foo has no polymorphic universes
so can't detect that way.
We could try passing the univ poly option down to schemes but a lot of
code change is needed.