[dune] Fix Dune build in Windows.
In order for Dune to work in Windows we need to tweak some script
calls, they need a POSIX shell and the `(run ...)` / `(system ...)`
actions use `cmd.exe` on Windows.

Hopefully, we will rely less on `bash` when Dune can understand Coq
libraries. This affects shell scripts in `kernel/**.sh` for example.

It is interesting to see how faster the Coq Windows build is with Dune
+ Windows.

There are some problems with PATHs that prevent the test suite from
working, these will be fixed in a future PR.