Change the index "p" to "s" in "type of branches".
In the description of destuctors,
"Type of branches" section uses "p" as the number of arguments.
It is confusing because "p" is used as the number of parameters.

After the section, "Typing rule." section uses "s" without definition as
I q1...qr t1...ts.
It seems that it is the number of arguments.

So, I changed "p" to "s".
"s" is also confusing with sorts, though.

(cherry picked from commit 7532f36176c2644bd9167c95fcd33ef3eff631af)