Change parameters p_1...p_r to q_1...q_r.
In the description of destructors,
"Type of branches" section and "Typing rule" section shares
the definition of "r" (and "s" from previous commit).

However actual parameters are p_1...p_r in the former section and
q_1...q_r in the latter section.

I guess it's because the latter section uses
p_1...p_l in other purpose: index of constructor for a inductive type.

So, I change the parameter p_1...p_r to q_1...q_r in the former section
to consistent with the latter section.