[save_proof] Make terminator API internal.
We refactor the terminator API to make it more internal. Indeed we
remove `set_terminator` and `get_terminator` is only there due to
access to internals in the STM `save_proof` path by the infamous
`?proof` parameter.

After this only 2 non-standard terminators remain: obligations and
derive. We will refactor those in next PRs.