[gitlab] [docker] [ci] Remove "edge" compiler switch.
Since a long time the compiler switch is not very useful as it is not
used to test any CI.

The `edge+flambda` version seems stable enough to carry out the `edge`
testing by itself, thus we remove the `egde` switch saving valuable
Docker image size and Gitlab runners.

We also tweak the `pkg:opam` job as to correctly set the version of
the pinned local package.