[funind] Apply automatic code formatting to the code base.
This is an experiment to format Coq's code base automatically.

I have chosen `funind` as indeed, even if our policy of no format-only
changes is IMO a good thing, the formatting in this part of the
codebase is so bad that it was making my editing there painful, thus
indeed such measure could be justified here.

The command chosen is:
```
ocamlformat --module-item-spacing=compact --inplace --break-sequences --sequence-style=terminator --cases-exp-indent=2 --field-space=loose *.ml *.mli
```

Indeed the result is not perfect but so bad either. I like it. Let me
know what think folks.