[coq_makefile] Enforce warn_error for plugins.
The amount of dangerous warnings in plugins is out of hand, including
some very serious ones.

As Coq developers are maintaining plugins these days it makes sense to
require the contributions to follow the same coding discipline as in
the main tree, thus we require the set of warnings fatal warnings to
be the same in Coq and in plugins.

We don't mark deprecation as fatal as to allow time for migration.

Fixes #6974

(cherry picked from commit 728216337e2541b37135d86c1b0206706cf4ed1a)