[dune] [refman] Generate documentation install files.
We generate a `coq-refman.install` file manually as Dune still doesn't
have install rules for directories.

We also allow appending to `coq.install` for the documentation of the
stdlib to be installed.

This should be solved upstream by adding support to:

- install rules for directories,
- native support for `coqdoc`.

however the current solution is useful for a few cases and doesn't
differ too much from what makefile actually does.