[interp] [numeral] Fix wrong unconditional call to VM reduction.
PR #8464 introduced an unconditional call to the VM in the
interpretation code.

This is wrong when the VM is not available [such as in JsCoq] and may
create problems.

I am not sure anyways that this PR is the right fix for the issue, it
is my impression that `Vnorm.cbv_vm` should never be called directly,
expert in the reduction API is needed here.