[ci] [windows] Use an install path requiring proper quoting.
Current quoting in Coq's infrastructure is not as systematic as it
should; IMHO this is a serious problem as it will make for random
failures would the user work with Coq inside a folder requiring
quoting.

This PR switches the CI system to use a destination directory
containing spaces; while this is still not optimal [the build dir is
free of spaces], it should provide some more assurance of proper
working wrt quoting for the runtime.