[engine] Remove kernel-level side effects push in refine.
This was noted by Gaëtan after some investigations, and indeed it
makes sense refine should have no business messing with this.

The functions creating the side-effects should be the ones that ensure
that the env is consistent.

I'm not fully sure how this is working yet tho, this part of the proof
engine is hard to follow.

Co-authored-by: Gaëtan Gilbert <gaetan.gilbert@skyskimmer.net>