[loadpath] Further cleanup after merge with MlTop.
We cleanup a bit the implementation of LoadPath which is not possible
as now all the loadpath logic is in the same place.

In particular, we remove exceptions in favor a `locate_result` monad.

More cleanup should still be possible, in particular
`locate_absolute_library` and `locate_qualified_library` should be
merged.